Removed test_empty_item_entered

+++ efill_app/views.py
@@ -4,7 +4,7 @@ from django.http import HttpResponse
 from efill_app.models import PersonalInfo
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
-
+import re
 # Create your views here.
 
 
@@ -13,10 +13,11 @@ def itr_form_view(request):
 	if request.method == 'POST':
 		try:
 			perosnal_obj = PersonalInfo.objects.create(pan_number=request.POST['pan_number'])
-			perosnal_obj.full_clean()
-		except ValidationError:
-			perosnal_obj.delete()
-			error = "Pan number cannot be empty."
+			if re.match(r'^[A-Z]{5}[0-9A-Z]{5}$',perosnal_obj.pan_number ):
+				perosnal_obj.full_clean()
+			else:
+				perosnal_obj.delete()
+				error = "Invalid PAN number"
 		except IntegrityError:
 			error = "Pan number should be unique"
